Moved PollingPredicate into seperate module for re-use.  Also removed other nasty Thread.sleeps
