+++ acceptance-tests/pom.xml
@@ -25,6 +25,11 @@
             <groupId>io.dropwizard</groupId>
             <artifactId>dropwizard-core</artifactId>
         </dependency>
+        <dependency>
+            <groupId>me.atam</groupId>
+            <artifactId>test-utils</artifactId>
+        </dependency>
+
     </dependencies>
