+++ acceptance-tests/src/test/java/me/atam/atam4jsampleapp/PassingTestAcceptanceTest.java
@@ -1,35 +1,38 @@
 package me.atam.atam4jsampleapp;
 
+import me.atam.atam4j.PollingPredicate;
+import me.atam.atam4j.dummytests.PassingTest;
 import me.atam.atam4j.health.AcceptanceTestsHealthCheck;
-import org.glassfish.jersey.client.JerseyClientBuilder;
 import org.junit.Test;
 
 import javax.ws.rs.core.Response;
+import java.util.function.Predicate;
+import java.util.function.Supplier;
 
 public class PassingTestAcceptanceTest extends AcceptanceTest{
 
 
+    public static final int MAX_ATTEMPTS = 2000;
+    public static final int RETRY_POLL_INTERVAL = 1;
+
     @Test
     public void givenSampleApplicationStartedWithPassingTest_whenHealthCheckCalledBeforeTestRun_thenTooEarlyMessageReceived(){
-        applicationConfigurationDropwizardTestSupport = Atam4JApplicationStarter.startApplicationWith(me.atam.atam4j.dummytests.PassingTest.class, 1000);
+        applicationConfigurationDropwizardTestSupport = Atam4JApplicationStarter.startApplicationWith(PassingTest.class, 1000);
         checkResponseIsOKAndWithMessage(AcceptanceTestsHealthCheck.TOO_EARLY_MESSAGE, getResponseFromHealthCheck());
     }
 
 
     @Test
     public void givenSampleApplicationStartedWithPassingTest_whenHealthCheckCalledAfterTestRUn_thenOKMessageReceived(){
-        applicationConfigurationDropwizardTestSupport = Atam4JApplicationStarter.startApplicationWith(me.atam.atam4j.dummytests.PassingTest.class);
-        waitForTwoMillis();
+        applicationConfigurationDropwizardTestSupport = Atam4JApplicationStarter.startApplicationWith(PassingTest.class, 0);
+
+        PollingPredicate<Response> responsePollingPredicate = new PollingPredicate<Response>(MAX_ATTEMPTS, RETRY_POLL_INTERVAL,
+                response -> response.readEntity(HealthCheckResult.class).getAcceptanceTestsHealthCheckResult().getMessage().equals(AcceptanceTestsHealthCheck.OK_MESSAGE),
+                () -> getResponseFromHealthCheck());
+        responsePollingPredicate.pollUntilPassedOrMaxAttemptsExceeded();
         checkResponseIsOKAndWithMessage(AcceptanceTestsHealthCheck.OK_MESSAGE, getResponseFromHealthCheck());
     }
 
-    private void waitForTwoMillis() {
-        try {
-            Thread.sleep(2);
-        } catch (InterruptedException e) {
-            throw new RuntimeException(e);
-        }
-    }
 
 
 }
