+++ test-utils/src/main/java/me/atam/atam4j/PollingPredicate.java
